Developed Splash Screen
Included the Resource for the Splash Screen as well as a rough mock up
for this weeks presentation. I found that we may run into issues
resizing the Splash Screen if we use a JLabel, since the image resource
isn't scaled to fit in the JLabel, rather the JLabel is scaled to the
resource I believe.